+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-page-views.js
@@ -112,11 +112,12 @@ window.GOVUK.analyticsGa4.analyticsModules = window.GOVUK.analyticsGa4.analytics
 
     getLanguage: function () {
       var content = document.getElementById('content')
+      var html = document.querySelector('html')
       if (content) {
         return content.getAttribute('lang') || this.nullValue
-      } else {
-        return this.nullValue
       }
+      // html.getAttribute('lang') is untested - Jasmine would not allow lang to be set on <html>.
+      return html.getAttribute('lang') || this.nullValue
     },
 
     getHistory: function () {