+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Fix bug with GA4 search results count ([PR #3594](https://github.com/alphagov/govuk_publishing_components/pull/3594))
+* Check HTML attribute when populating GA4 language ([PR #3598](https://github.com/alphagov/govuk_publishing_components/pull/3598))
 
 ## 35.15.2